+++ worlds/shapez_2/generate/events/operator_lines.py
@@ -21,10 +21,24 @@ def get_events(world: "Shapez2World",
 
     reg = regions["Events"]
     lines = world.operator_processors
+    all_unlocked = "Lock operator lines" not in world.options.location_modifiers
+    tab_unlocked = "Lock operator levels tab" not in world.options.location_modifiers
+
+    if tab_unlocked or "Operator Levels" in world.starting_items:
+        world.starting_items.append("[ACCESS] Operator levels")
+    else:
+        loc = Shapez2Location(
+            world.player, "[ACCESS] Operator levels", None, reg, LocationProgressType.PRIORITY,
+            lambda state: state.has("Operator Levels", world.player)
+        )
+        loc.place_locked_item(Shapez2Item(
+            "[ACCESS] Operator levels", ItemClassification.progression, None, world.player
+        ))
+        reg.locations.append(loc)
 
     for x in range(len(lines)):
         proc = tuple(event_by_processor[p] for p in lines[x])
-        if all(p in world.starting_items for p in proc):
+        if all_unlocked or all(p in world.starting_items for p in proc):
             world.starting_items.append(f"[ACCESS] Operator line {x+1}")
         else:
             loc = Shapez2Location(