+++ worlds/shapez_2/generate/items/mechanics.py
@@ -8,11 +8,17 @@ if TYPE_CHECKING:
 def generate_default(world: "Shapez2World") -> Iterator[Shapez2Item]:
     from ...data.items.mechanics import always
 
+    operator_levels_locked = "Lock operator levels tab" in world.options.location_modifiers
+
     for name, data in always.items():
-        yield Shapez2Item(name, data.classification(world), world.item_name_to_id[name], world.player)
+        if operator_levels_locked or name != "Operator Levels":
+            yield Shapez2Item(name, data.classification(world), world.item_name_to_id[name], world.player)
 
 
 def generate_starting(world: "Shapez2World") -> Iterator[str]:
     from ...data.items.mechanics import starting
+
+    if "Lock operator levels tab" not in world.options.location_modifiers:
+        yield "Operator Levels"
 
     yield from starting