+++ worlds/shapez_2/generate/items/operator_lines.py
@@ -8,7 +8,8 @@ if TYPE_CHECKING:
 def generate_default(world: "Shapez2World") -> Iterator[Shapez2Item]:
     from ...data.items.misc import operator_lines
 
-    for i in range(world.options.location_adjustments["Operator lines"]):
-        name = f"Operator line #{i + 1}"
-        data = operator_lines[name]
-        yield Shapez2Item(name, data.classification(world), world.item_name_to_id[name], world.player)
+    if "Lock operator lines" in world.options.location_modifiers:
+        for i in range(world.options.location_adjustments["Operator lines"]):
+            name = f"Operator line #{i + 1}"
+            data = operator_lines[name]
+            yield Shapez2Item(name, data.classification(world), world.item_name_to_id[name], world.player)